Moved CheckedListBox nested types to separate files
